Renamed `auth_token` to `authentication`
