+++ smelter-core/src/pipeline/utils/authentication.rs
