+++ smelter-core/src/pipeline/utils/mod.rs
@@ -1,4 +1,4 @@
-pub(crate) mod auth_token;
+pub(crate) mod authentication;
 pub(crate) mod channel;
 pub(crate) mod input_buffer;
